+++ nopassword/contrib/admin/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'nopassword.contrib.admin.apps.NoPasswordAdmin'