+++ nopassword/contrib/admin/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class NoPasswordAdmin(AppConfig):
+    name = 'nopassword.contrib.admin'
+    label = 'nopassword_admin'
+    verbose_name = "noPassword admin"