+++ nopassword/contrib/admin/views.py
@@ -1,0 +1,5 @@
+from nopassword.views import LoginView
+
+
+class AdminLoginView(LoginView):
+    template_name = 'nopassword_admin/login.html'