+++ nopassword/migrations/0002_remove_models.py
@@ -1,0 +1,18 @@
+from django.db import migrations
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('nopassword', '0001_initial'),
+    ]
+
+    operations = [
+        migrations.RemoveField(
+            model_name='logincode',
+            name='user',
+        ),
+        migrations.DeleteModel(
+            name='LoginCode',
+        ),
+    ]